Clean up PurchaseRatio
Removed the wholeEth variable, since it was duplicative. 

Could not find a way to programmatically set the number in 10*uint256(18) to whatever the user sets as the decimals. The issue is that the decimals is a uint8, which is not convertible into a uint256. Also tried a short if / else statement to handle this, but the uint8 variable didn't play nice with that either. Open to suggestions on how to set that "18" to be "6" etc. based on the user's choice of decimals.